[ws] improvements in logging and connection handling

+++ java/hiro-client/src/main/java/co/arago/hiro/client/api/WebSocketClient.java
@@ -11,7 +11,7 @@ public interface WebSocketClient {
    * @param body
    * @return the id of this request
    */
-  int sendMessage(String type, Map<String, String> headers, Map body);
+  long sendMessage(String type, Map<String, String> headers, Map body);
   
   void sendMessage(String message);
